requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
filelock==3.25.2
Flask==3.1.3
fonttools==4.59.1
fsspec==2026.2.0
gunicorn==25.1.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.1
kiwisolver==1.4.9
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
mysql==0.0.3
mysqlclient==2.2.7
networkx==3.6.1
numpy==2.3.1
openai==2.29.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.1
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==11.3.0
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.3
pypdfium2==5.6.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.3
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
selenium==4.29.0
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.3.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.6
wsproto==1.2.0
